+++ packages/database/src/outbox/entities/event-queue.entity.ts
@@ -1,7 +1,7 @@
 import type { EventPayload } from '../types/payload.registry.js';
 import { OutboxEntity } from './outbox.entity.js';
 import type { EventType } from '../types/event.type.js';
-import type { ServiceType } from '@volontariapp/shared';
+import type { Streams } from '@volontariapp/shared';
 
 export class EventQueueEntity<
   K extends EventType = EventType,
@@ -9,7 +9,7 @@ export class EventQueueEntity<
 > extends OutboxEntity<K> {
   version!: number;
 
-  targetServices: ServiceType[] = [];
+  targetServices: Streams[] = [];
 
   payload!: {
     before?: P;