+++ packages/shared/src/enums/index.ts
@@ -1,3 +1,4 @@
 export * from './tags-enum.js';
 export * from './user-roles-enum.js';
 export * from './service-type.enum.js';
+export * from './streams.enum.js';